:bug: Add game events to solve selection of placeholders
If you deleted a placeholder and a worker would start building it, the placeholder would still be selected. Instead of hardcoding the selection change in the placeholder (which it shouldn't do), I added a global event object that the interaction handler can listen to. We should probably design a better global event system, but without requirements it's hard to do something good.