🗃 Make having reports part of a component
The setup here is interesting. Because we can no longer rely on the "AddReport" event being subscribed to before reports are added, components need to support both sending that event, as well as responding to a query about all existing reports. This is something that could easily be done wrong, so I wanted to abstract that away. I decided to do so by fully hiding the event mechanics into an Aggregator helper class.